Make Results objects picklable.
Added __reduce__ and __setstate__ to make Results objects picklable.
Also added a save method to Results. This has to be used to save the
object with the same subclass. This means that if np.save is used we
loose the metadata added to the object as well as the additional methods
(such as plot) and when we load the file we get a standarda np.ndarray.